View all costs from the admin panel
