+++ frontend/vite.config.ts
@@ -43,6 +43,7 @@ export default defineConfig({
         './src/AdminPanel.tsx',
         './src/PartPanel.tsx',
         './src/CostTable.tsx',
+        './src/RateTable.tsx',
       ],
       output: [
         {
